minified error is over at last, no minification in server
