+++ spec/dummy/config/webpack/test.js
@@ -1,12 +1,10 @@
-process.env.NODE_ENV = process.env.NODE_ENV || 'development'
+process.env.NODE_ENV = process.env.NODE_ENV || 'test'
 
 const environment = require('./environment')
-const serverConfig = require('./server')
 const merge = require("webpack-merge")
 
-environment.splitChunks((config) => Object.assign({}, config, { optimization: { splitChunks: false }}))
-
 const clientEnvironment = merge(environment.toWebpackConfig(), {
+    mode: 'development',
     entry: {
         'vendor-bundle': [
             'jquery-ujs',
@@ -16,7 +14,12 @@ const clientEnvironment = merge(environment.toWebpackConfig(), {
         filename: '[name].js',
         chunkFilename: '[name].bundle.js',
         path: environment.config.output.path
-}
+    },
+    devtool: 'inline-source-map'
 })
+
+const serverConfig = require('./server')
+
+debugger
 
 module.exports = [clientEnvironment, serverConfig]
